+++ Runtime/ARFoundation/ARBackgroundRendererFeature.cs
@@ -160,14 +160,7 @@ namespace UnityEngine.XR.ARFoundation
                 cmd.BeginSample(k_CustomRenderPassName);
 
                 ARCameraBackground.AddBeforeBackgroundRenderHandler(cmd);
-                
-                // callback to schedule the release of the metal textures after rendering is complete
-                var callback = ARCameraBackground.NativeApi.Unity_Camera_GetTextureReleaseCallbackHandle();
-                if (callback != IntPtr.Zero)
-                {
-                    cmd.IssuePluginEvent(callback, 1);
-                }
-                
+
                 cmd.SetInvertCulling(m_InvertCulling);
 
                 cmd.SetViewProjectionMatrices(Matrix4x4.identity, projectionMatrix);